started fetch calls and stuff

+++ backend/db/migrate/002_create_distance_table.rb
@@ -2,7 +2,7 @@ class CreateDistanceTable < ActiveRecord::Migration[5.0]
     def change
         create_table :distances do |t|
             t.string :name
-            t.integer :lightyears
+            t.float :lightyears
             t.integer :combination_id
 
             t.timestamps
